+++ lib/api.dart
@@ -1,1 +1,8 @@
+/// API library
+///
+/// Provides API request scheduling and response handling
+///
+/// This library is designed to handle Mono API request rate limits.
+library api;
+
 export 'src/api.dart';