+++ lib/mcc.dart
@@ -1,1 +1,7 @@
+/// MCC library.
+///
+/// Provides MCC class to work with Merchant Category Codes that MCC dataset
+/// extensions can be applied to to provide descriptions and emojis.
+library mcc;
+
 export 'src/mcc/mcc.dart';