+++ lib/money.dart
@@ -1,1 +1,6 @@
+/// Money library
+///
+/// Provides Money class to work with monetary amounts
+library money;
+
 export 'src/money.dart';